save changes

+++ server/routes/users.js
@@ -1,9 +1,9 @@
 const express = require('express');
-
 const router = express.Router();
 const User = require('../models/user');
 const Sequelize = require("sequelize");
 const Op = Sequelize.Op;
+
 // list users 
 router.get('/', async (req,res)=>{
     const users = await User.findAll();
@@ -13,61 +13,68 @@ router.get('/', async (req,res)=>{
 
 // Create a user 
 router.post('/',async (req,res)=>{
-// validate the req first 
     const {customernumber,firstname,lastname,username,email,dateOfBirth,password,repeatPassword} = req.body;
-    const user = await User.create({
-        customernumber,
-        firstname,
-        lastname,
-        username,
-        email,
-        dateOfBirth,
-        password,
-        repeatPassword
-    });
-
-    res.send(user);
+    try {
+        const user = await User.create({
+            customernumber,
+            firstname,
+            lastname,
+            username,
+            email,
+            dateOfBirth,
+            password,
+            repeatPassword
+        });
+        res.send(user);
+    } catch (error) {
+            res.send(error);
+    }
 
 });
 
 // Update a user 
 router.put('/:id',async (req,res)=>{
     const {customernumber,firstname,lastname,username,email,dateOfBirth,password,repeatPassword} = req.body;
-    console.log(req.params.id);
-    const user = await User.findOne({where:{id: req.params.id }})
-    const updatedUser = await user.update({
-        customernumber,
-        firstname,
-        lastname,
-        username,
-        email,
-        dateOfBirth,
-        password,
-        repeatPassword
-    });
-
-    res.send(updatedUser);
-// also validate the request
-// verify existance
+    const user = await User.findOne({where:{id: req.params.id }});
+    if(!user) {
+        res.status(404).send({error : 'user not found'});
+    }else {
+        const updatedUser = await user.update({
+            customernumber,
+            firstname,
+            lastname,
+            username,
+            email,
+            dateOfBirth,
+            password,
+            repeatPassword
+        });
+    
+        res.send(updatedUser);
+    }
 });
 
 // get one user 
 router.get('/:id',async (req,res)=>{
     const user = await User.findOne({where:{id: req.params.id }});
-    console.log('calling get user');
-    res.status(200).send(user);
-// also validate the request
-// verify existance
+    if(!user) {
+        res.status(404).send({error : 'user not found'});
+    }else{
+        res.status(200).send(user);
+    }
+   
 });
 
 // Delete a user 
 router.delete('/:id', async (req,res)=> {
-    // verify existance 
     const user = await User.findOne({where:{id: req.params.id }});
-    const id = user.id;
-    user.destroy({force:true});
-    res.send({id});
-    
+    if(!user) {
+        res.status(404).send({error : 'user not found'});
+    }else {
+        const id = user.id;
+        user.destroy({force:true});
+        res.send({id});
+    }
 });
 
 // search users
